# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ApiErrorCode"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError",
        "type": "object"
      },
      "ApiErrorCode": {
        "enum": [
          "INVALID_REQUEST",
          "PROJECT_NOT_FOUND",
          "CHAPTER_NOT_FOUND",
          "UNSUPPORTED_IMPORT_TYPE",
          "MALFORMED_IMPORT",
          "MODEL_HARDWARE_UNSUPPORTED",
          "MODEL_DISK_SPACE_LOW",
          "MODEL_MANIFEST_FETCH_FAILED",
          "MODEL_MANIFEST_INVALID",
          "MODEL_DOWNLOAD_FAILED",
          "MODEL_CHECKSUM_MISMATCH"
        ],
        "title": "ApiErrorCode",
        "type": "string"
      },
      "Body_import_project_file_projects__project_id__imports_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_import_project_file_projects__project_id__imports_post",
        "type": "object"
      },
      "ChapterDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "createdAt": {
            "title": "Createdat",
            "type": "string"
          },
          "editorDoc": {
            "$ref": "#/components/schemas/ProseMirrorNode-Output"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "markdown": {
            "title": "Markdown",
            "type": "string"
          },
          "order": {
            "title": "Order",
            "type": "integer"
          },
          "projectId": {
            "title": "Projectid",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "sourceDocumentRecordId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcedocumentrecordid"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updatedAt": {
            "title": "Updatedat",
            "type": "string"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/ImportWarning"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "id",
          "projectId",
          "title",
          "order",
          "revision",
          "editorDoc",
          "markdown",
          "warnings",
          "createdAt",
          "updatedAt"
        ],
        "title": "ChapterDetailResponse",
        "type": "object"
      },
      "ChapterSummary": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "order": {
            "title": "Order",
            "type": "integer"
          },
          "sourceDocumentRecordId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcedocumentrecordid"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "warningCount": {
            "title": "Warningcount",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "title",
          "order",
          "warningCount"
        ],
        "title": "ChapterSummary",
        "type": "object"
      },
      "CreateImportResponse": {
        "additionalProperties": false,
        "properties": {
          "import": {
            "$ref": "#/components/schemas/ProjectImportSummary"
          },
          "project": {
            "$ref": "#/components/schemas/ProjectDetailResponse"
          }
        },
        "required": [
          "project",
          "import"
        ],
        "title": "CreateImportResponse",
        "type": "object"
      },
      "CreateProjectRequest": {
        "additionalProperties": false,
        "properties": {
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "CreateProjectRequest",
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImportState": {
        "enum": [
          "stored",
          "processing",
          "completed",
          "failed"
        ],
        "title": "ImportState",
        "type": "string"
      },
      "ImportWarning": {
        "additionalProperties": false,
        "properties": {
          "blockId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blockid"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "sourcePage": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcepage"
          }
        },
        "required": [
          "id",
          "code",
          "severity",
          "message"
        ],
        "title": "ImportWarning",
        "type": "object"
      },
      "ListProjectsResponse": {
        "additionalProperties": false,
        "properties": {
          "projects": {
            "items": {
              "$ref": "#/components/schemas/ProjectCard"
            },
            "title": "Projects",
            "type": "array"
          }
        },
        "required": [
          "projects"
        ],
        "title": "ListProjectsResponse",
        "type": "object"
      },
...
      "ProjectCard": {
        "additionalProperties": false,
        "properties": {
          "activeJobCount": {
            "title": "Activejobcount",
            "type": "integer"
          },
          "chapterCount": {
            "title": "Chaptercount",
            "type": "integer"
          },
          "createdAt": {
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "lastOpenedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastopenedat"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updatedAt": {
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "chapterCount",
          "lastOpenedAt",
          "activeJobCount",
          "createdAt",
          "updatedAt"
        ],
        "title": "ProjectCard",
        "type": "object"
      },
      "ProjectDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "chapters": {
            "items": {
              "$ref": "#/components/schemas/ChapterSummary"
            },
            "title": "Chapters",
            "type": "array"
          },
          "createdAt": {
            "title": "Createdat",
            "type": "string"
          },
          "defaultVoicePresetId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Defaultvoicepresetid"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "imports": {
            "items": {
              "$ref": "#/components/schemas/ProjectImportSummary"
            },
            "title": "Imports",
            "type": "array"
          },
          "lastOpenedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastopenedat"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updatedAt": {
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "chapters",
          "imports",
          "defaultVoicePresetId",
          "createdAt",
          "updatedAt",
          "lastOpenedAt"
        ],
        "title": "ProjectDetailResponse",
        "type": "object"
      },
      "ProjectImportSummary": {
        "additionalProperties": false,
        "properties": {
          "createdAt": {
            "title": "Createdat",
            "type": "string"
          },
          "failureMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failuremessage"
          },
          "fileSizeBytes": {
            "title": "Filesizebytes",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "sourceFileName": {
            "title": "Sourcefilename",
            "type": "string"
          },
          "sourceMimeType": {
            "title": "Sourcemimetype",
            "type": "string"
          },
          "sourceSha256": {
            "title": "Sourcesha256",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/ImportState"
          },
          "updatedAt": {
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "state",
          "sourceFileName",
          "sourceMimeType",
          "sourceSha256",
          "fileSizeBytes",
          "createdAt",
          "updatedAt",
          "failureMessage"
        ],
        "title": "ProjectImportSummary",
        "type": "object"
      },
      "ProseMirrorNode-Input": {
        "additionalProperties": false,
        "properties": {
          "attrs": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attrs"
          },
          "content": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ProseMirrorNode-Input"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type"
        ],
        "title": "ProseMirrorNode",
        "type": "object"
      },
      "ProseMirrorNode-Output": {
        "additionalProperties": false,
        "properties": {
          "attrs": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attrs"
          },
          "content": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ProseMirrorNode-Output"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type"
        ],
        "title": "ProseMirrorNode",
        "type": "object"
      },
...
      "UpdateChapterRequest": {
        "additionalProperties": false,
        "properties": {
          "editorDoc": {
            "$ref": "#/components/schemas/ProseMirrorNode-Input"
          }
        },
        "required": [
          "editorDoc"
        ],
        "title": "UpdateChapterRequest",
        "type": "object"
      },
      "UpdateProjectRequest": {
        "additionalProperties": false,
        "properties": {
          "defaultVoicePresetId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Defaultvoicepresetid"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "UpdateProjectRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Audaisy Runtime",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/projects": {
      "get": {
        "operationId": "list_projects_projects_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListProjectsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Projects",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "operationId": "create_project_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Create Project",
        "tags": [
          "projects"
        ]
      }
    },
    "/projects/{project_id}": {
      "delete": {
        "operationId": "delete_project_projects__project_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Project",
        "tags": [
          "projects"
        ]
      },
      "get": {
        "operationId": "get_project_projects__project_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Project",
        "tags": [
          "projects"
        ]
      },
      "patch": {
        "operationId": "patch_project_projects__project_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Patch Project",
        "tags": [
          "projects"
        ]
      }
    },
    "/projects/{project_id}/chapters/{chapter_id}": {
      "get": {
        "operationId": "get_chapter_projects__project_id__chapters__chapter_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "chapter_id",
            "required": true,
            "schema": {
              "title": "Chapter Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Chapter",
        "tags": [
          "chapters"
        ]
      },
      "patch": {
        "operationId": "patch_chapter_projects__project_id__chapters__chapter_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "chapter_id",
            "required": true,
            "schema": {
              "title": "Chapter Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChapterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Patch Chapter",
        "tags": [
          "chapters"
        ]
      }
    },
    "/projects/{project_id}/imports": {
      "post": {
        "operationId": "import_project_file_projects__project_id__imports_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_project_file_projects__project_id__imports_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateImportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Import Project File",
        "tags": [
          "projects"
        ]
      }
    },
...
  }
}